[#39] Flake8 update
This is maybe a bigger change then is wanted. I tried to to keep the
style of concatenation the same that is using `+`, but I couldn't find a
good way to do it that flake8 liked. Breaking on the `+` created a
message about a "line break before/after binary operator" and putting
the break else where was real ugly.

I also noticed that a number of files make use of f-strings and opted
for that for one of the examples so that it's consistent with the other
files and, I think, easier to read.